rename new voucher details page old page cleanup
