+++ react/assets/forus-webshop/scss/includes/blocks/block-transactions.scss
@@ -1,4 +1,4 @@
-.block.block-transactions-next {
+.block.block-transactions {
     display: flex;
     flex-direction: column;
