+++ react/assets/forus-webshop/scss/includes/blocks/block-voucher.scss
@@ -1,4 +1,4 @@
-.block.block-voucher-next {
+.block.block-voucher {
     display: flex;
     flex-direction: column;
     gap: 30px;
