+++ react/assets/forus-webshop/scss/webshop.scss
@@ -58,10 +58,10 @@
 @import 'includes/blocks/block-product';
 @import 'includes/blocks/record_categories';
 @import 'includes/blocks/records';
-@import 'includes/blocks/block-transactions-next';
+@import 'includes/blocks/block-transactions';
 @import 'includes/blocks/validations';
 @import 'includes/blocks/validators';
-@import 'includes/blocks/block-voucher-next';
+@import 'includes/blocks/block-voucher';
 @import 'includes/blocks/block-voucher-share-options';
 @import 'includes/blocks/vouchers';
 @import 'includes/blocks/fund-add-records';
