+++ react/src/webshop/components/pages/vouchers-show/hooks/useVoucherCard.tsx
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
-import Voucher from '../../../../../../dashboard/props/models/Voucher';
-import useComposeVoucherCardData from '../../../../../services/helpers/useComposeVoucherCardData';
+import Voucher from '../../../../../dashboard/props/models/Voucher';
+import useComposeVoucherCardData from '../../../../services/helpers/useComposeVoucherCardData';
 
 export default function useVoucherCard(voucher: Voucher) {
     const composeVoucherCardData = useComposeVoucherCardData();
