+++ react/src/webshop/router/routes.tsx
@@ -46,7 +46,7 @@ import FundsPreCheck from '../components/pages/funds-pre-check/FundsPreCheck';
 import ThrowError from '../components/pages_system/ThrowError';
 import Payouts from '../components/pages/payouts/Payouts';
 import Profile from '../components/pages/profile/Profile';
-import VouchersShowNext from '../components/pages/vouchers-show/VouchersShowNext';
+import VouchersShow from '../components/pages/vouchers-show/VouchersShow';
 import ProductsShow from '../components/pages/products-show/ProductsShow';
 
 const router = new RouterBuilder();
@@ -163,7 +163,7 @@ router.state('vouchers', <Vouchers />, {
     protected: true,
 });
 
-router.state('voucher', <VouchersShowNext />, {
+router.state('voucher', <VouchersShow />, {
     path: `/tegoeden/:number`,
     altPath: `/vouchers/:number`,
     protected: true,
